backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv>=1.0.0
python-multipart==0.0.9
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic[email]==2.6.1
pydantic-settings==2.1.0
email-validator>=2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.1,<4.1
faster-whisper>=1.0.0
chromadb==0.4.22
cryptography==42.0.2
httpx>=0.25.2,<0.26
ollama>=0.1.6
numpy==1.26.4
